Apply "formatting_changes" branch as a patch + revert most exclusively formatting changes + some extra changes

+++ src/game/behaviors/falling_pillar.inc.c
@@ -42,7 +42,7 @@ void bhv_falling_pillar_spawn_hitboxes(void) {
  * Computes the angle from current pillar position to 500 units in front of
  * Mario.
  */
-s16 bhv_falling_pillar_calculate_angle_in_front_of_mario(void) {
+s32 bhv_falling_pillar_calculate_angle_in_front_of_mario(void) {
     // Calculate target to be 500 units in front of Mario in
     // the direction he is facing (angle[1] is yaw).
     f32 targetX = sins(gMarioObject->header.gfx.angle[1]) * 500.0f + gMarioObject->header.gfx.pos[0];
